support pre-packed `Cmd` and `Pipeline` instance (#1)
Toolchain's internal async Redis client uses a MPMC channel to send requests to several async Tokio tasks that individually manage connections. Ownership of the request must be passed to the worker tasks. Thus, the async client pre-packs the data and writes the packed data into the channel. 

This PR allows those pre-packed requests to be provided to `redis::aio::ConnectionLike` implementations. The change intentionally is a hack that does not change the `redis::aio::ConnectionLike` trait.